src/v3doc/requirements.txt
numpy
openai
python-dotenv
rich
scipy
scikit-learn
tiktoken